--- src/main/java/com/toit/contents/attachments/AttachMentsController.java
@@ -1,16 +1,0 @@
-package com.toit.contents.attachments;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/items")
-@RequiredArgsConstructor
-public class AttachMentsController {
-    private final AttachMentsService attachMentsService;
-
-
-
-
-}
