+++ src/main/java/com/toit/view/pageitems/PageItemsUseCaseImpl.java
@@ -1,12 +1,12 @@
 package com.toit.view.pageitems;
 
-import com.toit.contents.attachments.AttachMentsService;
-import com.toit.contents.attachments.dto.response.AttachMentsFoldersImagesResponse;
-import com.toit.contents.attachments.dto.response.ItemsFoldersInFilesResponse;
-import com.toit.contents.links.LinksService;
-import com.toit.contents.links.dto.response.LinksGetInFoldersResponse;
-import com.toit.contents.texts.TextsService;
-import com.toit.contents.texts.dto.response.TextsGetInFoldersResponse;
+import com.toit.items.attachments.AttachMentsService;
+import com.toit.items.attachments.dto.response.AttachMentsImagesGetInFoldersResponse;
+import com.toit.items.attachments.dto.response.AttachMentsFilesGetInFoldersResponse;
+import com.toit.items.links.LinksService;
+import com.toit.items.links.dto.response.LinksGetInFoldersResponse;
+import com.toit.items.texts.TextsService;
+import com.toit.items.texts.dto.response.TextsGetInFoldersResponse;
 import com.toit.view.pageitems.dto.response.PageFoldersInItemsAllResponse;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -24,8 +24,8 @@ public class PageItemsUseCaseImpl implements PageItemsUseCase{
     public PageFoldersInItemsAllResponse getOnFoldersInItemsAll(Long usersId, Long foldersId) {
         List<LinksGetInFoldersResponse> links = linksService.getLinksInFolders(usersId, foldersId);
         List<TextsGetInFoldersResponse> texts = textsService.getTextsInFolders(usersId, foldersId);
-        List<ItemsFoldersInFilesResponse> files = attachMentsService.getFoldersFiles(usersId, foldersId);
-        List<AttachMentsFoldersImagesResponse> images = attachMentsService.getFoldersImages(usersId, foldersId);
+        List<AttachMentsFilesGetInFoldersResponse> files = attachMentsService.getFilesInFolders(usersId, foldersId);
+        List<AttachMentsImagesGetInFoldersResponse> images = attachMentsService.getImagesInFolders(usersId, foldersId);
 
         return new PageFoldersInItemsAllResponse(usersId, foldersId, links, texts, files, images);
     }
